trees.cpp - 16/10/2017
Added an abstract comparable base class, used to allow for a more generic BST class, with a tree that can store Comparables instead of integers. 

Needed to use 'dynamic_cast' to tell the compiler that, although the object was passed as a Comparable data type, it is actually a type derived from it.

Used some operator overloading for assignment.