stack.cpp 17/10/17
stack.push(data) wasn't working because I was making a larger array but then forgetting to set the last element of the array to the data input. Works now.

stack.push(data) is very inefficient. It must copy the entire stack into a new area of memory that is one element larger.

Very fast to read and pop(), but not fast to push. pop() function doesn't delete the memory location of the top element, it just decrements the 'length' field.