Add files via upload
Header file for matrix of complex numbers code. Had to change some lines from the original matrix file. In the operator overloading functions of multiplication, addition and subtraction the rhs before the assignment operator must be cast as the template type. I think this is because there is more than one operation, and the result may be saved in a temporary memory location?