generate_strategies - 20/10/17
Added in a vector of char vectors, that store essentially bool data types (but vector<bool> doesn't work). These are flags which tell the script whether a variable has been printed so that it isn't printed twice in a line. In hindsight, it needn't be a vector of vectors, just a 1D array will do. You don't need to store this info for previous lines, only the current line.